restructure so there is only one BodyView script. BodyDataConverter and BodyDataReceiver now implement a common interface. Fix size issue by having skeletons inherit position from parent BodyView game object. Fixed the removal problem in receiver when skeleton leaves scene by added a boolean to the message that indicates when a body has been removed on the server side.
